% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{current_mode_data}
\alias{current_mode_data}
\title{current_mode function}
\usage{
current_mode_data(
  data_type,
  date_range = NULL,
  custom_polygon = NULL,
  meteolanddb,
  geometry_id,
  progress_obj,
  lang
)
}
\description{
check and retrieve the data for the current mode
}
