% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_data}
\alias{get_data}
\title{get data function}
\usage{
get_data(
  data_type,
  data_mode,
  date_range = NULL,
  custom_polygon = NULL,
  meteolanddb,
  geometry_id,
  progress_obj,
  lang
)
}
\description{
check and retrieve the data for the selected mode and type
}
