% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cvOutput.R
\name{mod_cvOutput}
\alias{mod_cvOutput}
\alias{mod_cv}
\title{mod_cvOutput and mod_cv}
\usage{
mod_cvOutput(id)

mod_cv(input, output, session, meteolanddb, lang)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{lang}{lang selected}

\item{meteolandb}{object to access the meteolanddb}
}
\description{
Shiny module to show the corssvalidations table
}
\details{
mod_cv generates the crossvalidations table
}
